coding style correct
