--- dsPIC30F/I2Cdev/I2CdevdsPic30F.X/nbproject/Makefile-genesis.properties
@@ -1,9 +1,0 @@
-#
-#Wed Apr 26 14:40:24 CST 2017
-default.com-microchip-mplab-nbide-toolchainXC16-XC16LanguageToolchain.md5=ebaa75370c1463d650408e4144ba15de
-default.languagetoolchain.dir=/opt/microchip/xc16/v1.30/bin
-configurations-xml=e23099a44127c530cab293bfc5554ec5
-com-microchip-mplab-nbide-embedded-makeproject-MakeProject.md5=8b140fb87818dc96ed4f726b716b40e9
-default.languagetoolchain.version=1.30
-host.platform=linux
-conf.ids=default
