--- dsPIC30F/I2Cdev/I2CdevdsPic30F.X/nbproject/Makefile-variables.mk
@@ -1,13 +1,0 @@
-#
-# Generated - do not edit!
-#
-# NOCDDL
-#
-CND_BASEDIR=`pwd`
-# default configuration
-CND_ARTIFACT_DIR_default=dist/default/production
-CND_ARTIFACT_NAME_default=I2CdevdsPic30F.X.a
-CND_ARTIFACT_PATH_default=dist/default/production/I2CdevdsPic30F.X.a
-CND_PACKAGE_DIR_default=${CND_DISTDIR}/default/package
-CND_PACKAGE_NAME_default=I2CdevdsPic30F.X.tar
-CND_PACKAGE_PATH_default=${CND_DISTDIR}/default/package/I2CdevdsPic30F.X.tar
