--- dsPIC30F/MPU6050/Examples/MPU6050_raw.X/funclist
@@ -1,37 +1,0 @@
-_MPU6050_initialize: CODE, 6334 0 24
-_I2Cdev_writeBytes: CODE, 5202 0 156
-_MPU6050_getDeviceID: CODE, 6310 0 24
-__stringdata: SMALLCONST, 3851 0 193
-_putch: CODE, 6552 0 12
-_MPU6050_getXGyroOffset: CODE, 6188 0 42
-_MPU6050_getYGyroOffset: CODE, 6104 0 42
-_MPU6050_getZGyroOffset: CODE, 6020 0 42
-_dpowers: SMALLCONST, 3841 0 10
-_I2Cdev_readByte: CODE, 6514 0 20
-_I2Cdev_writeBit: CODE, 5474 0 110
-_main: CODE, 4046 0 342
-_I2Cdev_readBits: CODE, 5584 0 108
-_MPU6050_setFullScaleGyroRange: CODE, 6358 0 24
-_MPU6050_setSleepEnabled: CODE, 6474 0 20
-_MPU6050: CODE, 6584 0 8
-_MPU6050_setClockSource: CODE, 6406 0 24
-__initialization: CODE, 6430 0 12
-___lwdiv: CODE, 5884 0 74
-_MPU6050_setFullScaleAccelRange: CODE, 6382 0 24
-_baud1USART: CODE, 6576 0 8
-_Write1USART: CODE, 6534 0 18
-_printf: CODE, 4706 0 294
-___lwmod: CODE, 5958 0 62
-_MPU6050_testConnection: CODE, 6564 0 12
-_ReadI2C1: CODE, 6272 0 38
-_MPU6050_getXAccelOffset: CODE, 6230 0 42
-_MPU6050_getYAccelOffset: CODE, 6146 0 42
-_MPU6050_getZAccelOffset: CODE, 6062 0 42
-_WriteI2C1: CODE, 5692 0 108
-_I2Cdev_writeByte: CODE, 6494 0 20
-_Open1USART: CODE, 5800 0 84
-_I2Cdev_readBytes: CODE, 4388 0 318
-_I2Cdev_writeBits: CODE, 5358 0 116
-_MPU6050_getMotion6: CODE, 5000 0 202
-_OpenI2C1: CODE, 6452 0 22
-Total: 2739
