--- dsPIC30F/MPU6050/Examples/MPU6050_raw.X/nbproject/Makefile-genesis.properties
@@ -1,8 +1,0 @@
-#
-#Fri Nov 28 22:55:53 CET 2014
-default.languagetoolchain.dir=C\:\\Program Files (x86)\\Microchip\\xc8\\v1.32\\bin
-com-microchip-mplab-nbide-embedded-makeproject-MakeProject.md5=20b996ec0591c93034157100c3d90ac3
-default.languagetoolchain.version=1.32
-host.platform=windows
-conf.ids=default
-default.com-microchip-mplab-nbide-toolchainXC8-XC8LanguageToolchain.md5=533e5c34133c5d768bcb87ffcec948e1
