--- dsPIC30F/MPU6050/Examples/MPU6050_raw.X/nbproject/Makefile-variables.mk
@@ -1,13 +1,0 @@
-#
-# Generated - do not edit!
-#
-# NOCDDL
-#
-CND_BASEDIR=`pwd`
-# default configuration
-CND_ARTIFACT_DIR_default=dist/default/production
-CND_ARTIFACT_NAME_default=MPU6050_raw.X.production.hex
-CND_ARTIFACT_PATH_default=dist/default/production/MPU6050_raw.X.production.hex
-CND_PACKAGE_DIR_default=${CND_DISTDIR}/default/package
-CND_PACKAGE_NAME_default=mpu6050raw.x.tar
-CND_PACKAGE_PATH_default=${CND_DISTDIR}/default/package/mpu6050raw.x.tar
