--- dsPIC30F/MPU6050/Examples/MPU6050_raw.X/nbproject/project.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><project xmlns="http://www.netbeans.org/ns/project/1">
-    <type>com.microchip.mplab.nbide.embedded.makeproject</type>
-    <configuration>
-        <data xmlns="http://www.netbeans.org/ns/make-project/1">
-            <name>MPU6050_raw</name>
-            <creation-uuid>84892e89-df8e-441e-9542-2524d6b2dc5a</creation-uuid>
-            <make-project-type>0</make-project-type>
-            <c-extensions>c</c-extensions>
-            <cpp-extensions/>
-            <header-extensions>h</header-extensions>
-            <asminc-extensions/>
-            <sourceEncoding>UTF-8</sourceEncoding>
-            <make-dep-projects/>
-        </data>
-    </configuration>
-</project>
